prepare repo, service and routing

+++ server/src/main/kotlin/domain/repo/UserRepository.kt
@@ -6,7 +6,7 @@ import com.moashrafff.domain.model.request.RegisterRequest
 
 interface UserRepository {
     suspend fun createUser(registerRequest: RegisterRequest): User?
-    suspend fun getUserById(id: Int): User?
+    suspend fun getUserById(id: Long): User?
     suspend fun login(request: LoginRequest): User?
     fun hashPassword(password: String): String
     fun verifyPassword(plainPassword: String, hashedPassword: String): Boolean
